+++ supertokens_python/recipe/dashboard/api/__init__.py
@@ -11,8 +11,10 @@
 # WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 # License for the specific language governing permissions and limitations
 # under the License.
+from .analytics import handle_analytics_post
 from .api_key_protector import api_key_protector
 from .dashboard import handle_dashboard_api
+from .search.getTags import handle_get_tags
 from .signin import handle_emailpassword_signin_api
 from .signout import handle_emailpassword_signout_api
 from .userdetails.user_delete import handle_user_delete
@@ -49,4 +51,6 @@ __all__ = [
     "handle_email_verify_token_post",
     "handle_emailpassword_signin_api",
     "handle_emailpassword_signout_api",
+    "handle_get_tags",
+    "handle_analytics_post",
 ]