+++ supertokens_python/recipe/dashboard/api/implementation.py
@@ -20,6 +20,8 @@ from typing import TYPE_CHECKING, Any, Dict
 from supertokens_python import Supertokens
 from supertokens_python.normalised_url_domain import NormalisedURLDomain
 from supertokens_python.normalised_url_path import NormalisedURLPath
+from supertokens_python.querier import Querier
+from supertokens_python.utils import is_version_gte
 
 from ..constants import DASHBOARD_API
 from ..interfaces import APIInterface
@@ -54,6 +56,14 @@ class APIImplementation(APIInterface):
                 NormalisedURLPath(DASHBOARD_API)
             ).get_as_string_dangerous()
 
+            is_search_enabled: bool = False
+            querier = Querier.get_instance(options.recipe_id)
+            cdiVersion = await querier.get_api_version()
+            if not cdiVersion:
+                is_search_enabled = True
+            elif is_version_gte(cdiVersion, "2.20"):
+                is_search_enabled = True
+
             return Template(
                 dedent(
                     """
@@ -65,6 +75,7 @@ class APIImplementation(APIInterface):
                             window.dashboardAppPath = "${dashboardPath}"
                             window.connectionURI = "${connectionURI}"
                             window.authMode = "${authMode}"
+                            window.isSearchEnabled = "${isSearchEnabled}"
                         </script>
                         <script defer src="${bundleDomain}/static/js/bundle.js"></script></head>
                         <link href="${bundleDomain}/static/css/main.css" rel="stylesheet" type="text/css">
@@ -82,6 +93,7 @@ class APIImplementation(APIInterface):
                 dashboardPath=dashboard_path,
                 connectionURI=connection_uri,
                 authMode=auth_mode,
+                isSearchEnabled=str(is_search_enabled).lower(),
             )
 
         self.dashboard_get = dashboard_get