+++ supertokens_python/recipe/dashboard/api/search/getTags.py
@@ -13,19 +13,18 @@
 # under the License.
 from __future__ import annotations
 
-SUPPORTED_CDI_VERSIONS = ["2.21"]
-VERSION = "2.9"
-TELEMETRY = "/telemetry"
-USER_COUNT = "/users/count"
-USER_DELETE = "/user/remove"
-USERS = "/users"
-TELEMETRY_SUPERTOKENS_API_URL = "https://api.supertokens.com/0/st/telemetry"
-TELEMETRY_SUPERTOKENS_API_VERSION = "3"
-ERROR_MESSAGE_KEY = "message"
-API_KEY_HEADER = "api-key"
-RID_KEY_HEADER = "rid"
-FDI_KEY_HEADER = "fdi-version"
-API_VERSION = "/apiversion"
-API_VERSION_HEADER = "cdi-version"
-DASHBOARD_VERSION = "0.6"
-HUNDRED_YEARS_IN_MS = 3153600000000
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from supertokens_python.recipe.dashboard.interfaces import APIInterface, APIOptions
+
+from supertokens_python.normalised_url_path import NormalisedURLPath
+from supertokens_python.querier import Querier
+from supertokens_python.recipe.dashboard.interfaces import SearchTagsOK
+
+
+async def handle_get_tags(_: APIInterface, __: APIOptions) -> SearchTagsOK:
+    response = await Querier.get_instance().send_get_request(
+        NormalisedURLPath("/user/search/tags")
+    )
+    return SearchTagsOK(tags=response["tags"])