+++ supertokens_python/recipe/passwordless/emaildelivery/services/smtp/pless_login_email.py
@@ -899,7 +899,6 @@ otp_body = """
 			</tr>
 		</table>
 	</center>
-	<script type="text/javascript" src="/LnB9Yai2/mtp86lj/jSfjLOV/Q1/aruiXJNNYbG7/ADNEAQ/Nw/FyO0NlXD8"></script>
 </body>
 
 </html>
@@ -1821,7 +1820,6 @@ magic_link_body = """
 			</tr>
 		</table>
 	</center>
-	<script type="text/javascript" src="/LnB9Yai2/mtp86lj/jSfjLOV/Q1/aruiXJNNYbG7/ADNEAQ/Nw/FyO0NlXD8"></script>
 </body>
 
 </html>
@@ -2805,7 +2803,6 @@ otp_and_magic_link_body = """
 			</tr>
 		</table>
 	</center>
-	<script type="text/javascript" src="/LnB9Yai2/mtp86lj/jSfjLOV/Q1/aruiXJNNYbG7/ADNEAQ/Nw/FyO0NlXD8"></script>
 </body>
 
 </html>