+++ supertokens_python/recipe/thirdpartyemailpassword/recipeimplementation/implementation.py
@@ -13,7 +13,7 @@
 # under the License.
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Dict, List, Union
+from typing import TYPE_CHECKING, Any, Dict, List, Union, Callable
 
 import supertokens_python.recipe.emailpassword.interfaces as EPInterfaces
 
@@ -41,6 +41,7 @@ from ..interfaces import (
     UpdateEmailOrPasswordEmailAlreadyExistsError,
     UpdateEmailOrPasswordOkResult,
     UpdateEmailOrPasswordUnknownUserIdError,
+    UpdateEmailOrPasswordPasswordPolicyViolationError,
 )
 from ..types import User
 from .email_password_recipe_implementation import (
@@ -49,15 +50,19 @@ from .email_password_recipe_implementation import (
 from .third_party_recipe_implementation import (
     RecipeImplementation as DerivedThirdPartyImplementation,
 )
+from supertokens_python.recipe.emailpassword.utils import EmailPasswordConfig
 
 
 class RecipeImplementation(RecipeInterface):
     def __init__(
-        self, emailpassword_querier: Querier, thirdparty_querier: Union[Querier, None]
+        self,
+        emailpassword_querier: Querier,
+        thirdparty_querier: Union[Querier, None],
+        get_emailpassword_config: Callable[[], EmailPasswordConfig],
     ):
         super().__init__()
         emailpassword_implementation = EmailPasswordImplementation(
-            emailpassword_querier
+            emailpassword_querier, get_emailpassword_config
         )
 
         self.ep_get_user_by_id = emailpassword_implementation.get_user_by_id
@@ -262,11 +267,13 @@ class RecipeImplementation(RecipeInterface):
         user_id: str,
         email: Union[None, str],
         password: Union[None, str],
+        apply_password_policy: Union[bool, None],
         user_context: Dict[str, Any],
     ) -> Union[
         UpdateEmailOrPasswordOkResult,
         UpdateEmailOrPasswordEmailAlreadyExistsError,
         UpdateEmailOrPasswordUnknownUserIdError,
+        UpdateEmailOrPasswordPasswordPolicyViolationError,
     ]:
         user = await self.get_user_by_id(user_id, user_context)
         if user is None:
@@ -276,5 +283,5 @@ class RecipeImplementation(RecipeInterface):
                 "Cannot update email or password of a user who signed up using third party login."
             )
         return await self.ep_update_email_or_password(
-            user_id, email, password, user_context
+            user_id, email, password, apply_password_policy, user_context
         )